Remove tests for now; license

+++ include/errno.h
@@ -1,14 +1,32 @@
+/*
+ *  Copyright (C) 2019 James Larrowe
+ *
+ *  This file is part of Minilibc.
+ *
+ *  Minilibc is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU Lesser General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  Minilibc is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU Lesser General Public License for more details.
+ *
+ *  You should have received a copy of the GNU Lesser General Public License
+ *  along with Minilibc.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 #ifndef _ERRNO_H
 #define _ERRNO_H 1
 
-#include <stdint.h>
 #include <features.h>
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-typedef intptr_t errno_t;
+typedef int errno_t;
 
 #define errno (*__get_errno())
 
@@ -54,3 +72,4 @@ errno_t errno;
 #endif
 
 #endif
+
